--- a/app/clubs/viewsets.py
+++ b/app/clubs/viewsets.py
@@ -1,3 +1,11 @@
+from core.abstracts.viewsets import (
+    CustomLimitOffsetPagination,
+    FilterBackendBase,
+    ModelViewSetBase,
+    ObjectViewDetailsPermissions,
+    ViewSetBase,
+)
+from core.models import Major
 from django.db.models import Count, Prefetch
 from django.shortcuts import get_object_or_404
 from django.utils.decorators import method_decorator
@@ -9,6 +17,7 @@ from rest_framework import exceptions, filters, mixins, permissions, status
 from rest_framework.decorators import action
 from rest_framework.generics import GenericAPIView
 from rest_framework.response import Response
+from users.models import User
 
 from clubs.models import (
     Club,
@@ -37,15 +46,6 @@ from clubs.serializers import (
     TeamSerializer,
 )
 from clubs.services import ClubService
-from core.abstracts.viewsets import (
-    CustomLimitOffsetPagination,
-    FilterBackendBase,
-    ModelViewSetBase,
-    ObjectViewDetailsPermissions,
-    ViewSetBase,
-)
-from core.models import Major
-from users.models import User
 
 
 def get_user_club_or_404(club_id: int, user: User):
@@ -311,6 +311,18 @@ class TeamViewSet(ClubNestedViewSetBase):
             ),
         )
     )
+
+    # Cache detail view for 2 hours (per Authorization header)
+    @method_decorator(cache_page(60 * 60 * 2))
+    @method_decorator(vary_on_headers("Authorization"))
+    def retrieve(self, request, *args, **kwargs):
+        return super().retrieve(request, *args, **kwargs)
+
+    # Cache list view for 2 hours (per Authorization header)
+    @method_decorator(cache_page(60 * 60 * 2))
+    @method_decorator(vary_on_headers("Authorization"))
+    def list(self, request, *args, **kwargs):
+        return super().list(request, *args, **kwargs)
 
 
 class InviteClubMemberView(GenericAPIView):