--- a/app/lib/allauth.py
+++ b/app/lib/allauth.py
@@ -3,6 +3,8 @@ from allauth.headless.adapter import DefaultHeadlessAdapter
 from allauth.socialaccount.adapter import DefaultSocialAccountAdapter
 from allauth.socialaccount.providers.base import Provider
 from allauth.socialaccount.providers.oauth2.provider import OAuth2Provider
+from rest_framework import exceptions
+from users.models import User
 
 ProviderType = Provider
 OauthProviderType = OAuth2Provider
@@ -17,7 +19,17 @@ class CustomAccountAdapter(DefaultAccountAdapter):
 
 
 class CustomSocialAccountAdapter(DefaultSocialAccountAdapter):
-    pass
+
+    def populate_user(self, request, sociallogin, data):
+        email = data.get("email")
+
+        existing_user = User.objects.find_by_email(email=email)
+        if existing_user:
+            raise exceptions.AuthenticationFailed(
+                detail=f"User already exists with email {email}"
+            )
+
+        return super().populate_user(request, sociallogin, data)
 
 
 __all__ = ["ProviderType", "OauthProviderType"]