--- a/app/users/apis.py
+++ b/app/users/apis.py
@@ -28,7 +28,11 @@ urlpatterns = [
     path("me/", viewsets.ManageUserView.as_view(), name="me"),
     path("oauth-directory/", viewsets.OauthDirectoryView.as_view()),
     path("me/providers/", viewsets.SocialProviderViewSet.as_view({"get": "list"})),
-    path("oauth/provider/", viewsets.RedirectToProviderView.as_view()),
+    path(
+        "oauth/provider/",
+        viewsets.RedirectToProviderView.as_view({'post': 'post'}),
+        name="oauth_redirect",
+    ),
     path(
         "oauth/provider/redirect/",
         viewsets.ReturnFromOauthView.as_view(),