--- a/app/users/apis.py
+++ b/app/users/apis.py
@@ -30,7 +30,7 @@ urlpatterns = [
     path("me/providers/", viewsets.SocialProviderViewSet.as_view({"get": "list"})),
     path(
         "oauth/provider/",
-        viewsets.RedirectToProviderView.as_view({'post': 'post'}),
+        viewsets.RedirectToProviderView.as_view({"post": "post"}),
         name="oauth_redirect",
     ),
     path(