--- a/app/users/utils.py
+++ b/app/users/utils.py
@@ -1,0 +1,13 @@
+from app.settings import SCHOOL_EMAIL_DOMAIN
+from django.core import exceptions
+from django.core.validators import validate_email
+
+
+def is_school_email(email: str):
+    """Check if the supplied email address is a valid school email."""
+
+    try:
+        validate_email(email)
+        return email.endswith(SCHOOL_EMAIL_DOMAIN)
+    except exceptions.ValidationError:
+        return False