--- a/deploy/proxy/default.conf.tpl
+++ b/deploy/proxy/default.conf.tpl
@@ -1,7 +1,13 @@
 gzip on;
 
+upstream app_upstream {
+  server app:9000;
+}
+
 server {
   listen 8080;
+  
+  resolver 127.0.0.11 valid=5s;
   
   location /static {
     alias /vol/web;
@@ -17,7 +23,9 @@ server {
   }
   
   location / {
-    uwsgi_pass              "$SERVER_URI";
+    # uwsgi_pass              "$SERVER_URI";
+    # uwsgi_pass http://app_upstream;
+    proxy_pass http://app_upstream;
     
     proxy_set_header        Host "$host";
     proxy_set_header        X-Forwarded-For "$proxy_add_x_forwarded_for";