--- a/sample.env
+++ b/sample.env
@@ -68,3 +68,6 @@ S3_STORAGE_BUCKET_REGION=""
 
 PROXY_DOCS_URI="http://localhost:8080/api/v1/docs"
 PROXY_DELAY='0s'
+
+APP_REPLICAS=3  # Docker-compose worker count
+WORKER_COUNT=20  # uWSGI worker count per replica