--- a/scripts/healthcheck.sh
+++ b/scripts/healthcheck.sh
@@ -1,0 +1,6 @@
+#!/bin/sh
+
+set -e
+
+echo "Checking health for http://localhost:${PORT}..."
+curl -f http://localhost:${PORT}