### Description

- Allow merging of user accounts
- Merge user accounts when one user verifies email of another user
- Add "merge users" action in django admin
- Created unit tests for oauth
- Improve docker compose for prod
- Add service in prod compose file that replicates app containers
- Create new uwsgi config ini file (so we can leave comments next to each setting)
- Add caching for teams
- Add two new env vars to tune performance on production machine: `APP_REPLICAS`, `WORKER_COUNT`

### Performance metrics

Test command:
```sh
loadtest -n 1000 -c 50 --rps 50 http://localhost:8080/api/v1/club/club-previews/\?limit\=10000
```

NOTE: Caching was disabled for club previews when testing, then reenabled afterwards

Results before:
- 0 errors
- longest request: 47,964 ms
- mean latency: 19,095 ms
- total time: 58s
- effective rps: 17

Results after (20 processes, 5 replicas):
- 0 errors
- longest request: 28,823 ms (-19,141 ms)
- mean latency: 15,402 ms (-3,639 ms)
- total time: 48s (-10s)
- effective rps: 21 (+4 requests per second)